back/auth-microservice/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.28.0
SQLAlchemy==2.0.28
python-decouple==3.8
psycopg2==2.9.9